Refactor: Redis Entity, Repository 패키지 구조 이동
